refactor: treat singlefile torrents as multifile
