+++ torrentfile/tracker_test.go
@@ -26,7 +26,7 @@ func TestBuildTrackerURL(t *testing.T) {
 	peerID := utils.BTString{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20}
 	const port uint16 = 6882
 
-	url, err := torrent.buildTrackerURL(peerID, port)
+	url, err := torrent.BuildTrackerURL(peerID, port)
 	expected := "http://bttracker.debian.org:6969/announce?compact=1&downloaded=0&info_hash=%D8%F79%CE%C3%28%95l%CC%5B%BF%1F%86%D9%FD%CF%DB%A8%CE%B6&left=351272960&peer_id=%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%0E%0F%10%11%12%13%14&port=6882&uploaded=0"
 
 	if err != nil {
@@ -73,7 +73,7 @@ func TestRequestPeers(t *testing.T) {
 		{IP: net.IP{127, 0, 0, 1}, Port: 6889},
 	}
 
-	peers, err := tf.requestPeers(peerID, port)
+	peers, err := tf.RequestPeers(peerID, port)
 
 	assert.Nil(t, err)
 	assert.Equal(t, expected, peers)
